Added support for legged systems in geometric system plotter
Added support for quasistatic legged systems,

The scope of the added system/ code is explained below:

Changes are made to: "sysf_two_pin_legged_rigidbody.m"

System geometry: ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
The body is assumed to a be a single rigid link and all legs originate from the center of the body, hence they move around a single disk. In the the main sysf code (refer above this for the name of the file), we consider only two legs.

How to make changes:
Single disk legged systems are supported where the transform from the body-frame to the leg-tip is just a function of the hip angle. If number of legs need to changed add more link lengths to the left and right sides of the s.geometry.linklengths array (note that the center link corresponds to the body so that the array always has odd length and even number of legs), and make corresponding changes to s.physics.st_friction_coeff array.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

System Physics:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
We have modeled this in such a way that only one leg can be pinned to the ground. This forces us to obtain the local connection in a form different from standard geometric mechanics connections -- the translational velocity correspond to the body frame and the rotational velocity corresponds to the pinned-leg.
Secondly, the system's dissipation metric is obtained from isotropic viscous friction effects of the non-pinned leg. Currently, the solver doesn't not support other dissipation metrics such as 'anisotropy' and 'inertial'.

How to make changes:
Change the 's.physics.st_friction_coeff' array to match the length of the 's.geometry.linklengths' array, and make sure only one entry is 'Inf' since that corresponds to the pinned leg. Every other leg should have '0' and the rigid body link at the center should have a value of nan since it is not contacting the ground.
Artificial change to the viscous friction dissipation metric, by changing the 'nu' coefficient values in the function that computes the metric (you can make it anisotropic, etc manually). Support for some of these functions will be available in a little bit.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The files that compute the connections, etc can be found 'output.dependency' path argument in the beginning of the sysf file.